requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.4
colorama==0.4.6
coloredlogs==15.0.1
cryptography==41.0.1
decorator==4.4.2
dnspython==2.4.1
exceptiongroup==1.1.1
fastapi==0.99.0
ffmpeg-python==0.2.0
flatbuffers==23.5.26
frozenlist==1.3.3
future==0.18.3
h11==0.14.0
httptools==0.5.0
humanfriendly==10.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
lazy_loader==0.3
lit==16.0.6
llvmlite==0.40.1
lxml==4.9.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numba==0.57.1
numpy==1.24.4
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
onnxruntime==1.15.1
openai==0.27.8
opencv-python-headless==4.8.0.74
packaging==23.1
pdfdocx==1.6
pdfminer.six==20221105
Pillow==10.0.0
platformdirs==3.8.0
pooch==1.7.0
proglog==0.1.10
protobuf==4.23.3
pycparser==2.21
pydantic==1.10.10
PyMatting==1.1.8
pymongo==4.4.1
PyMuPDF==1.19.2
python-dateutil==2.8.2
python-docx==0.8.10
python-dotenv==1.0.0
python-multipart==0.0.6
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.6.3
rembg==2.0.49
requests==2.31.0
scikit-image==0.21.0
scipy==1.11.1
six==1.16.0
sniffio==1.3.0
SpeechRecognition==3.10.0
starlette==0.27.0
sympy==1.12
tifffile==2023.7.4
tqdm==4.65.0
typing_extensions==4.7.0
urllib3==2.0.3
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
youtube-dl==2021.12.17
